Check if daily filenames exist before backing up

+++ database-backup.sh
@@ -19,12 +19,18 @@ do
 	project=${sites__project[$n]}
 	domain=${sites__domain[$n]}
 	path=${sites__path[$n]}
+	filename=${date_full}-${project}
 
-	echo "Exporting ${project} at ${domain}"
-	wp db export "${date_full}-${project}.sql" --ssh="${hostname}" --path="${path}"
-	ssh $hostname "rm *.sql.gz"
-	ssh $hostname "gzip ${date_full}-${project}.sql"
-	scp $hostname:"${date_full}-${project}.sql.gz" "./${date_full}-db-files/"
+	if [ -f "./${date_full}-db-files/${filename}.sql.gz" ];
+	then
+		echo "${filename} has already been backed up today."
+	else
+		echo "Exporting ${project} at ${domain}"
+		wp db export "${date_full}-${project}.sql" --ssh="${hostname}" --path="${path}"
+		ssh $hostname "rm *.sql.gz"
+		ssh $hostname "gzip ${date_full}-${project}.sql"
+		scp $hostname:"${date_full}-${project}.sql.gz" "./${date_full}-db-files/"
+	fi
 done
 
 # Synchronize DB backup with B2.
